requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1
huggingface_hub==0.19.4
Pillow==10.1.0
langchain 
langchain-community
langchain-ollama
python-dotenv
